add Route entity and some security feaches

+++ src/main/java/by/bsuir/timetable/repository/StationRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 public interface StationRepository extends MongoRepository<Station, String> {
     Station findByName(String name);
 
-    Station findByCode(Integer code);
+    Station findByCode(Long code);
 }
